requirements.txt
phidata==2.7.2
google-generativeai==0.8.3
streamlit==1.40.2
tensorflow